00_RESTful URL setting

+++ boards/views.py
@@ -1,8 +1,19 @@
 from django.shortcuts import render, redirect
 from .models import Board
+# from pprint import pprint
 
 # Create your views here.
 def index(request):
+    # pprint(request)
+    # pprint(type(request))
+    # pprint(dir(request))
+    # pprint(request.scheme)
+    # pprint(request.get_host())
+    # pprint(request.get_full_path())
+    # pprint(request.build_absolute_uri())
+    # pprint(request.META)
+    # pprint(request.method)
+    
     boards = Board.objects.order_by('-pk')
     context = {
         'boards' : boards,
@@ -10,15 +21,24 @@ def index(request):
     return render(request, 'boards/index.html', context)
     
 def new(request):
-    return render(request, 'boards/new.html')
+    if request.method == 'POST':
+        # CREATE
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        
+        board = Board(title=title, content=content)
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        return render(request, 'boards/new.html')
     
-def create(request):
-    title = request.POST.get('title')
-    content = request.POST.get('content')
+# def create(request):
+#     title = request.POST.get('title')
+#     content = request.POST.get('content')
     
-    board = Board(title=title, content=content)
-    board.save()
-    return redirect('boards:detail', board.pk)
+#     board = Board(title=title, content=content)
+#     board.save()
+#     return redirect('boards:detail', board.pk)
     
 def detail(request, pk):
     board = Board.objects.get(pk=pk)
@@ -29,20 +49,32 @@ def detail(request, pk):
     
 def delete(request, pk):
     board = Board.objects.get(pk=pk)
-    board.delete()
-    return redirect('boards:index')
+    if request.method == 'POST':
+        board.delete()
+        return redirect('boards:index')
+    else:
+        return redirect('boards:detail', board.pk)
     
 def edit(request, pk):
-    board = Board.objects.get(pk=pk)
-    context = {
-        'board' : board,
-    }
-    return render(request, 'boards/edit.html', context)
+    if request.method == 'POST':
+        # UPDATE
+        board = Board.objects.get(pk=pk)
+        board.title = request.POST.get('title')
+        board.content = request.POST.get('content')
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        # EDIT
+        board = Board.objects.get(pk=pk)
+        context = {
+            'board' : board,
+        }
+        return render(request, 'boards/edit.html', context)
     
-def update(request, pk):
-    board = Board.objects.get(pk=pk)
-    board.title = request.POST.get('title')
-    board.content = request.POST.get('content')
-    board.save()
-    return redirect('boards:detail', board.pk)
+# def update(request, pk):
+#     board = Board.objects.get(pk=pk)
+#     board.title = request.POST.get('title')
+#     board.content = request.POST.get('content')
+#     board.save()
+#     return redirect('boards:detail', board.pk)
